requirements.txt
lxml>=4.5.2
Mastodon.py>=1.5.1
opencv-python>=4.4.0.42
pandas>=1.1.0
Pillow>=7.2.0
plotly>=4.9.0
psutil>=5.7.2
socketIO-client-nexus>=0.7.6
wikipedia>=1.4.0
xmltodict>=0.12.0
jaconv
cairosvg
transformers
fugashi
ipadic
unidic-lite
# deepsparse
sentencepiece
onnxruntime
openai
pydantic-settings
